Create run chart with highlighted improvements where applicable

This will plot the original dataframe, with highlighted runs of improvement. It will also return a dataframe showing the improvment data

runcharter(df, med_rows = 13, runlength = 9, chart_title = NULL,
  chart_subtitle = NULL, direction = "below", faceted = TRUE,
  facet_cols = NULL, save_plot = FALSE, plot_extension = "png", ...)

Arguments

df
dataframe containing columns "date", "y", "grp"

med_rows
the number of rows you wish to base the initial median calculation over

runlength
how long a run of consecutive points should be before re-basing the median

chart_title
title for the final chart

chart_subtitle
subtitle for chart

direction
look for runs "below" or "above" the median

faceted
if you dont need a faceted / trellis display set this to FALSE

facet_cols
the number of columns required for a faceted plot. Ignored if faceted is set to FALSE

save_plot
should the plot be saved? Calls ggsave on the last plot, saving in the current directory, if TRUE.

plot_extension
one of "png","pdf" or other valid extension for saving ggplot2 plots. Used in the call to ggsave.

...
further arguments passed on to function

Value

runchart and a dataframe showing sustained run data if appropriate

Examples

runcharter(data, med_rows = 13, runlength = 9, chart_title = "Automated runs analysis",
direction = "above", faceted = FALSE)

#> Error in UseMethod("arrange_"): no applicable method for 'arrange_' applied to an object of class "function"